split large components into separate files
